ref(utils): Stop setting `transaction` in `requestDataIntegration`
This is not really necessary anymore - it only sets this on transaction events, and those get the `transaction` in different places already anyhow.

With this, we can also actually remove some other stuff. One method is exported from utils but not otherwise used, we can also drop this in v9.

Finally, this was also the only place that used `route` on the request, so we can also get rid of this in `remix`, which is weird anyhow because we set it for errors there but don't even use it for them.